Revert "Split infractions into another plugin, add some confirm options"
This reverts commit 19006a1785f68cd95c65b942264d439ab111e307.
