+++ config.example.yaml
@@ -12,7 +12,6 @@ bot:
     - rowboat.plugins.core
     - rowboat.plugins.modlog
     - rowboat.plugins.admin
-    - rowboat.plugins.infractions
     - rowboat.plugins.spam
     - rowboat.plugins.censor
     - rowboat.plugins.reddit
